+++ javascript/src/next.ts
@@ -300,25 +300,57 @@ export function splice<T>(
   del: number,
   newText?: string,
 ) {
+  const objPath = absoluteObjPath(doc, path, "splice")
   if (!_is_proxy(doc)) {
     throw new RangeError("object cannot be modified outside of a change block")
   }
   const state = _state(doc, false)
-  const objectId = _obj(doc)
-  if (!objectId) {
-    throw new RangeError("invalid object for splice")
-  }
   _clear_cache(doc)
 
-  path.unshift(objectId)
-  const value = path.join("/")
-
-  index = cursorToIndex(state, value, index)
+  index = cursorToIndex(state, objPath, index)
 
   try {
-    return state.handle.splice(value, index, del, newText)
+    return state.handle.splice(objPath, index, del, newText)
   } catch (e) {
     throw new RangeError(`Cannot splice: ${e}`)
+  }
+}
+
+/**
+ * Update the value of a string
+ *
+ * @typeParam T - The type of the value contained in the document
+ * @param doc - The document to modify
+ * @param path - The path to the string to modify
+ * @param newText - The new text to update the value to
+ *
+ * @remarks
+ * This will calculate a diff between the current value and the new value and
+ * then convert that diff into calls to {@link splice}. This will produce results
+ * which don't merge as well as directly capturing the user input actions, but
+ * sometimes it's not possible to capture user input and this is the best you
+ * can do.
+ *
+ * This is an experimental API and may change in the future.
+ *
+ * @beta
+ */
+export function updateText(
+  doc: Doc<unknown>,
+  path: stable.Prop[],
+  newText: string,
+) {
+  const objPath = absoluteObjPath(doc, path, "updateText")
+  if (!_is_proxy(doc)) {
+    throw new RangeError("object cannot be modified outside of a change block")
+  }
+  const state = _state(doc, false)
+  _clear_cache(doc)
+
+  try {
+    return state.handle.updateText(objPath, newText)
+  } catch (e) {
+    throw new RangeError(`Cannot updateText: ${e}`)
   }
 }
 
@@ -345,17 +377,11 @@ export function getCursor<T>(
   path: stable.Prop[],
   index: number,
 ): Cursor {
+  const objPath = absoluteObjPath(doc, path, "getCursor")
   const state = _state(doc, false)
-  const objectId = _obj(doc)
-  if (!objectId) {
-    throw new RangeError("invalid object for getCursor")
-  }
-
-  path.unshift(objectId)
-  const value = path.join("/")
 
   try {
-    return state.handle.getCursor(value, index)
+    return state.handle.getCursor(objPath, index)
   } catch (e) {
     throw new RangeError(`Cannot getCursor: ${e}`)
   }
@@ -375,17 +401,11 @@ export function getCursorPosition<T>(
   path: stable.Prop[],
   cursor: Cursor,
 ): number {
+  const objPath = absoluteObjPath(doc, path, "getCursorPosition")
   const state = _state(doc, false)
-  const objectId = _obj(doc)
-  if (!objectId) {
-    throw new RangeError("invalid object for getCursorPosition")
-  }
-
-  path.unshift(objectId)
-  const value = path.join("/")
 
   try {
-    return state.handle.getCursorPosition(value, cursor)
+    return state.handle.getCursorPosition(objPath, cursor)
   } catch (e) {
     throw new RangeError(`Cannot getCursorPosition: ${e}`)
   }
@@ -398,20 +418,14 @@ export function mark<T>(
   name: string,
   value: MarkValue,
 ) {
+  const objPath = absoluteObjPath(doc, path, "mark")
   if (!_is_proxy(doc)) {
     throw new RangeError("object cannot be modified outside of a change block")
   }
   const state = _state(doc, false)
-  const objectId = _obj(doc)
-  if (!objectId) {
-    throw new RangeError("invalid object for mark")
-  }
-
-  path.unshift(objectId)
-  const obj = path.join("/")
 
   try {
-    return state.handle.mark(obj, range, name, value)
+    return state.handle.mark(objPath, range, name, value)
   } catch (e) {
     throw new RangeError(`Cannot mark: ${e}`)
   }
@@ -423,35 +437,25 @@ export function unmark<T>(
   range: MarkRange,
   name: string,
 ) {
+  const objPath = absoluteObjPath(doc, path, "unmark")
   if (!_is_proxy(doc)) {
     throw new RangeError("object cannot be modified outside of a change block")
   }
   const state = _state(doc, false)
-  const objectId = _obj(doc)
-  if (!objectId) {
-    throw new RangeError("invalid object for unmark")
-  }
-
-  path.unshift(objectId)
-  const obj = path.join("/")
 
   try {
-    return state.handle.unmark(obj, range, name)
+    return state.handle.unmark(objPath, range, name)
   } catch (e) {
     throw new RangeError(`Cannot unmark: ${e}`)
   }
 }
 
 export function marks<T>(doc: Doc<T>, path: stable.Prop[]): Mark[] {
+  const objPath = absoluteObjPath(doc, path, "marks")
   const state = _state(doc, false)
-  const objectId = _obj(doc)
-  if (!objectId) {
-    throw new RangeError("invalid object for marks")
-  }
-  path.unshift(objectId)
-  const obj = path.join("/")
+
   try {
-    return state.handle.marks(obj)
+    return state.handle.marks(objPath)
   } catch (e) {
     throw new RangeError(`Cannot call marks(): ${e}`)
   }
@@ -462,15 +466,10 @@ export function marksAt<T>(
   path: stable.Prop[],
   index: number,
 ): MarkSet {
+  const objPath = absoluteObjPath(doc, path, "marksAt")
   const state = _state(doc, false)
-  const objectId = _obj(doc)
-  if (!objectId) {
-    throw new RangeError("invalid object for marksAt")
-  }
-  path.unshift(objectId)
-  const obj = path.join("/")
   try {
-    return state.handle.marksAt(obj, index)
+    return state.handle.marksAt(objPath, index)
   } catch (e) {
     throw new RangeError(`Cannot call marksAt(): ${e}`)
   }
@@ -534,4 +533,18 @@ export function getConflicts<T>(
   } else {
     return undefined
   }
+}
+
+function absoluteObjPath(
+  doc: Doc<unknown>,
+  path: stable.Prop[],
+  functionName: string,
+): string {
+  path = path.slice()
+  const objectId = _obj(doc)
+  if (!objectId) {
+    throw new RangeError(`invalid object for ${functionName}`)
+  }
+  path.unshift(objectId)
+  return path.join("/")
 }