+++ rust/automerge/src/lib.rs
@@ -281,6 +281,7 @@ mod read;
 mod sequence_tree;
 mod storage;
 pub mod sync;
+mod text_diff;
 mod text_value;
 pub mod transaction;
 mod types;