+++ rust/automerge/src/query/nth.rs
@@ -112,7 +112,7 @@ impl<'a> TreeQuery<'a> for Nth<'a> {
             QueryResult::Finish
         } else {
             if let Some(m) = self.marks.as_mut() {
-                m.process(op)
+                m.process(op, self.clock.as_ref())
             }
             let visible = op.visible_at(self.clock.as_ref());
             let key = op.elemid_or_key();