+++ rust/CHANGELOG.md
@@ -1,3 +1,24 @@
+# 0.5.8
+
+* Fix a bug where the logic to rollback a transaction on error could panic
+* Fix a bug where marks were calculated incorrectly when viewing a document at
+  a particular set of heads (i.e. not the "current" heads)
+* Update the `LoadOptions::migrate_strings` logic to no-op if there are no
+  strings to convert
+
+# 0.5.7
+
+* Update itertools dependency to 0.12.0
+* Fix a bug in `Read::get_marks` which caused it to ignore any heads passed to
+  it and always return the latest marks (only relevant if you ever passed 
+  `Some(_)` as the heads argument of `Read::get_marks(objid, index, heads)`)
+
+# 0.5.6
+
+* Add `Transactable::update_text`, which calculates a diff between the current
+  value of a text field and a new value and converts that diff into a set of 
+  splice operations
+
 # 0.5.5
 
 * Fix a sync protocol backwards compatibility gotcha which caused 0.5.4 peers